optimize createsession queries, fix files Delete, fix php 8.1, add misc tests, fix bugs

+++ Andromeda/Core/Database/BaseObject.php
@@ -260,7 +260,8 @@ abstract class BaseObject
     /** Deletes the object from the database */
     public function Delete() : void
     {
-        $this->database->DeleteObject($this);
+        if (!$this->isDeleted)
+            $this->database->DeleteObject($this);
     }
     
     /**
